# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RuknAuto Invoice Extraction API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface InvoiceItem {
  id?: number;
  /** @nullable */
  partNumber?: string | null;
  description: string;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  unitCost: number;
  total?: number;
  memoryMatch?: boolean;
  /** @nullable */
  memoryConfidence?: number | null;
  needsManualInput?: boolean;
  /** @nullable */
  originalPartNumber?: string | null;
  packFactor?: number;
}

export interface Invoice {
  id: number;
  /** @nullable */
  invoiceNumber?: string | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  date?: string | null;
  status: string;
  /** @nullable */
  imageBase64?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  itemCount?: number | null;
  createdAt: string;
  items?: InvoiceItem[];
}

export interface ExtractInvoiceInput {
  imageBase64: string;
  mimeType: string;
  model?: string;
}

export interface ExtractedInvoice {
  invoiceId: number;
  /** @nullable */
  invoiceNumber?: string | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  date?: string | null;
  items: InvoiceItem[];
}

export interface SaveInvoiceInput {
  invoiceNumber?: string;
  supplier?: string;
  date?: string;
  items: InvoiceItem[];
}

export interface Part {
  id: number;
  description: string;
  partNumber: string;
  /** @nullable */
  originalPartNumber?: string | null;
  packFactor?: number;
  usageCount: number;
  createdAt: string;
}

export interface PartInput {
  description: string;
  partNumber: string;
  /** @nullable */
  originalPartNumber?: string | null;
  packFactor?: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}

export interface Stats {
  totalInvoices: number;
  totalParts: number;
  totalItemsExtracted: number;
  successRate: number;
  recentActivity: ActivityItem[];
}